# Improve performance of zip