Late in the process of #15356, I discovered that forcing inlining can greatly improve the performance of `zip`. The `zipsame` approach developed in that PR remains attractive, but I need a 3-iterator version, and this should be good enough for me to proceed with ReshapedArrays (which is my real goal here).

Benchmarks:
```jl
function iter1!(A, B, C)
    @assert size(A) == size(B) == size(C)
    for I in eachindex(A, B, C)
        @inbounds A[I] = B[I]*C[I]
    end
    A
end

function iter3!(A, B, C)
    @assert length(A) == length(B) == length(C)
    for (IA, IB, IC) in zip(eachindex(A), eachindex(B), eachindex(C))
        @inbounds A[IA] = B[IB]*C[IC]
    end
    A
end

sz = (100,100,100,100);
rng = map(n->1:n, sz)
B = sub(rand(sz...), rng...)
C = sub(rand(sz...), rng...)
A = similar(B)
iter1!(A, B, C)
iter3!(A, B, C)
@time iter1!(A, B, C)
@time iter3!(A, B, C)
nothing
```
Results:

master:
```
  0.631552 seconds (197 allocations: 11.855 KB)
  2.735151 seconds (4 allocations: 160 bytes)
```

This PR:
```
  0.630054 seconds (159 allocations: 11.137 KB)
  0.796275 seconds (4 allocations: 160 bytes)
```
So a 3-4 fold improvement.