@nalimilan 2016-03-06
Have you checked whether all of these annotations are needed, or are only some of these functions missing inlining?