@timholy 2016-03-06
For any particular example, I would bet one can find cases that don't need inlining. For example, with two `LinearFast` arrays you don't need any of them, but make them `LinearSlow` and suddenly you do. It's quite possible that `start` could be omitted, but it doesn't seem worth taking the risk, especially when I get around to implementing a 3-iterator version of `ZipSame`. Inlining is basically giving LLVM to reason about the code, something that's not possible if it doesn't inline.