@nalimilan 2016-03-06
Fair enough. I don't think there are cases in which we wouldn't want these to be inlined anyway.